requirements.txt
langchain==0.0.279
pinecone_client==2.2.2
streamlit==1.26.0
unstructured
unstructured[pdf]
openai
chromadb
tiktoken
pyserial